# .tmp/capturas/openapi.json
{"openapi":"3.1.0","info":{"title":"TechPC API - FastAPI","description":"API REST del proyecto TechPC (cuarto avance). Autenticación JWT, hashing de contraseñas, control de roles y operaciones CRUD sobre la base de datos SQL.","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/auth/login":{"post":{"tags":["Autenticación"],"summary":"Login","description":"Inicio de sesión: valida correo y contraseña y retorna un JWT.","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["Autenticación"],"summary":"Get Me","description":"Perfil del usuario autenticado (requiere token).","operationId":"get_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioOut"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/auth/roles":{"get":{"tags":["Autenticación"],"summary":"Get Roles","description":"Lista de roles del sistema (requiere autenticación).","operationId":"get_roles_api_auth_roles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/usuarios/registro":{"post":{"tags":["Usuarios"],"summary":"Registrar Usuario","description":"Registro de un cliente nuevo: valida datos, verifica duplicados\ny almacena la contraseña con hashing seguro (bcrypt).","operationId":"registrar_usuario_api_usuarios_registro_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioRegistro"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios":{"get":{"tags":["Usuarios"],"summary":"Listar Usuarios","description":"Lista todos los usuarios (solo administrador).","operationId":"listar_usuarios_api_usuarios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Listar Usuarios Api Usuarios Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/usuarios/{usuario_id}":{"get":{"tags":["Usuarios"],"summary":"Obtener Usuario","description":"Consulta un usuario por ID (solo administrador).","operationId":"obtener_usuario_api_usuarios__usuario_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Obtener Usuario Api Usuarios  Usuario Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Usuarios"],"summary":"Actualizar Usuario","description":"Actualiza los datos de un usuario (solo administrador).","operationId":"actualizar_usuario_api_usuarios__usuario_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Usuarios"],"summary":"Eliminar Usuario","description":"Elimina un usuario (solo administrador).","operationId":"eliminar_usuario_api_usuarios__usuario_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}/estado":{"patch":{"tags":["Usuarios"],"summary":"Cambiar Estado Usuario","description":"Cambia el estado de un usuario a activo/inactivo (solo administrador).","operationId":"cambiar_estado_usuario_api_usuarios__usuario_id__estado_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstadoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/productos":{"get":{"tags":["Productos"],"summary":"Listar Productos","description":"Lista todos los productos con su categoría.","operationId":"listar_productos_api_productos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Productos"],"summary":"Crear Producto","description":"Crea un producto nuevo (solo administrador o empleado).","operationId":"crear_producto_api_productos_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/productos/categorias":{"get":{"tags":["Productos"],"summary":"Listar Categorias","description":"Lista las categorías de productos.","operationId":"listar_categorias_api_productos_categorias_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/productos/{producto_id}":{"get":{"tags":["Productos"],"summary":"Obtener Producto","description":"Consulta un producto por ID.","operationId":"obtener_producto_api_productos__producto_id__get","parameters":[{"name":"producto_id","in":"path","required":true,"schema":{"type":"integer","title":"Producto Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Productos"],"summary":"Actualizar Producto","description":"Actualiza un producto (solo administrador o empleado).","operationId":"actualizar_producto_api_productos__producto_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"producto_id","in":"path","required":true,"schema":{"type":"integer","title":"Producto Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Productos"],"summary":"Eliminar Producto","description":"Elimina un producto (solo administrador).","operationId":"eliminar_producto_api_productos__producto_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"producto_id","in":"path","required":true,"schema":{"type":"integer","title":"Producto Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/servicios":{"get":{"tags":["Servicios"],"summary":"Listar Servicios","description":"Lista todos los servicios.","operationId":"listar_servicios_api_servicios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Servicios"],"summary":"Crear Servicio","description":"Crea un servicio nuevo (solo administrador o empleado).","operationId":"crear_servicio_api_servicios_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicioCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/servicios/{servicio_id}":{"get":{"tags":["Servicios"],"summary":"Obtener Servicio","description":"Consulta un servicio por ID.","operationId":"obtener_servicio_api_servicios__servicio_id__get","parameters":[{"name":"servicio_id","in":"path","required":true,"schema":{"type":"integer","title":"Servicio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Servicios"],"summary":"Actualizar Servicio","description":"Actualiza un servicio (solo administrador o empleado).","operationId":"actualizar_servicio_api_servicios__servicio_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"servicio_id","in":"path","required":true,"schema":{"type":"integer","title":"Servicio Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServicioUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Servicios"],"summary":"Eliminar Servicio","description":"Elimina un servicio (solo administrador).","operationId":"eliminar_servicio_api_servicios__servicio_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"servicio_id","in":"path","required":true,"schema":{"type":"integer","title":"Servicio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EstadoUpdate":{"properties":{"estado":{"type":"string","pattern":"^(activo|inactivo)$","title":"Estado"}},"type":"object","required":["estado"],"title":"EstadoUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"contrasena":{"type":"string","maxLength":255,"minLength":1,"title":"Contrasena"}},"type":"object","required":["email","contrasena"],"title":"LoginRequest"},"LoginResponse":{"properties":{"message":{"type":"string","title":"Message"},"token":{"type":"string","title":"Token"},"user":{"additionalProperties":true,"type":"object","title":"User"}},"type":"object","required":["message","token","user"],"title":"LoginResponse"},"ProductoCreate":{"properties":{"nombre":{"type":"string","maxLength":100,"minLength":2,"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Descripcion"},"precio":{"type":"number","maximum":999999999999.99,"exclusiveMinimum":0.0,"title":"Precio"},"stock":{"type":"integer","maximum":1000000.0,"minimum":0.0,"title":"Stock"},"imagen_url":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Imagen Url"},"categoria_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Categoria Id"}},"type":"object","required":["nombre","precio","stock"],"title":"ProductoCreate"},"ProductoUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Descripcion"},"precio":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Precio"},"stock":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Stock"},"imagen_url":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Imagen Url"},"categoria_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Categoria Id"},"estado":{"anyOf":[{"type":"string","pattern":"^(activo|inactivo)$"},{"type":"null"}],"title":"Estado"}},"type":"object","title":"ProductoUpdate"},"ServicioCreate":{"properties":{"nombre":{"type":"string","maxLength":100,"minLength":2,"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Descripcion"},"precio":{"type":"number","maximum":999999999999.99,"exclusiveMinimum":0.0,"title":"Precio"},"duracion_estimada":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Duracion Estimada"},"imagen_url":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Imagen Url"}},"type":"object","required":["nombre","precio"],"title":"ServicioCreate"},"ServicioUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string","maxLength":100,"minLength":2},{"type":"null"}],"title":"Nombre"},"descripcion":{"anyOf":[{"type":"string","maxLength":1000},{"type":"null"}],"title":"Descripcion"},"precio":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Precio"},"duracion_estimada":{"anyOf":[{"type":"string","maxLength":50},{"type":"null"}],"title":"Duracion Estimada"},"imagen_url":{"anyOf":[{"type":"string","maxLength":255},{"type":"null"}],"title":"Imagen Url"},"estado":{"anyOf":[{"type":"string","pattern":"^(activo|inactivo)$"},{"type":"null"}],"title":"Estado"}},"type":"object","title":"ServicioUpdate"},"UsuarioOut":{"properties":{"id":{"type":"integer","title":"Id"},"nombre":{"type":"string","title":"Nombre"},"apellido":{"type":"string","title":"Apellido"},"tipo_documento":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tipo Documento"},"numero_documento":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Numero Documento"},"direccion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Direccion"},"telefono":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Telefono"},"correo":{"type":"string","title":"Correo"},"estado":{"type":"string","title":"Estado"},"rol_id":{"type":"integer","title":"Rol Id"},"rol_nombre":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rol Nombre"},"fecha_registro":{"anyOf":[{},{"type":"null"}],"title":"Fecha Registro"}},"type":"object","required":["id","nombre","apellido","correo","estado","rol_id"],"title":"UsuarioOut","description":"Representación de un usuario sin contraseña."},"UsuarioRegistro":{"properties":{"nombre":{"type":"string","maxLength":50,"minLength":2,"title":"Nombre"},"apellido":{"type":"string","maxLength":50,"minLength":2,"title":"Apellido"},"tipoDocumento":{"type":"string","maxLength":5,"minLength":2,"title":"Tipodocumento"},"numeroDocumento":{"type":"string","maxLength":12,"minLength":6,"title":"Numerodocumento"},"direccion":{"type":"string","maxLength":80,"minLength":5,"title":"Direccion"},"telefono":{"type":"string","maxLength":16,"minLength":7,"title":"Telefono"},"correo":{"type":"string","format":"email","title":"Correo"},"contrasena":{"type":"string","maxLength":255,"minLength":8,"title":"Contrasena"}},"type":"object","required":["nombre","apellido","tipoDocumento","numeroDocumento","direccion","telefono","correo","contrasena"],"title":"UsuarioRegistro","description":"El Frontend envía tipoDocumento/numeroDocumento (camelCase);\ntambién se aceptan los nombres de columna en snake_case."},"UsuarioUpdate":{"properties":{"nombre":{"anyOf":[{"type":"string","maxLength":50,"minLength":2},{"type":"null"}],"title":"Nombre"},"apellido":{"anyOf":[{"type":"string","maxLength":50,"minLength":2},{"type":"null"}],"title":"Apellido"},"tipo_documento":{"anyOf":[{"type":"string","maxLength":5,"minLength":2},{"type":"null"}],"title":"Tipo Documento"},"numero_documento":{"anyOf":[{"type":"string","maxLength":12,"minLength":6},{"type":"null"}],"title":"Numero Documento"},"direccion":{"anyOf":[{"type":"string","maxLength":80,"minLength":5},{"type":"null"}],"title":"Direccion"},"telefono":{"anyOf":[{"type":"string","maxLength":16,"minLength":7},{"type":"null"}],"title":"Telefono"},"correo":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Correo"},"contrasena":{"anyOf":[{"type":"string","maxLength":255,"minLength":8},{"type":"null"}],"title":"Contrasena"},"rol_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rol Id"},"estado":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Estado"}},"type":"object","title":"UsuarioUpdate","description":"Todos los campos opcionales para la edición de usuarios."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}